Rebalancing TNG
We find a map that satisfies all constraints (number of replication
relationships among nodes, equal number of vbuckets replicated per
master/slave pair, equal number of vbuckets at each position in the
replication on each node) while attempting to minimize the number of
changes.

The new algorithm is to generate all of the replication chains we want
in the final map, then map vbuckets to those chains based on the
number of moves required to change the vbucket config from the old
chain to the new one.

Change-Id: I4709ac6f2575923816fc08173288a044aeb5d14e
Reviewed-on: http://review.membase.org/4941
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>