return config tuple thats compatible with 1.6.5. MB-3554
In 1.6.5.1 we made config saving async, but this also changed format
of config record. We're passing this record around as part of config
replication. ns_config_rep:do_pull calls ns_config:config_dynamic on
result of remote ns_config:get and this fails if config records have
different versions.

This change makes sure that we always return "old" version of config
tuple. Rest of the change ensures that we're handling this reduced
config tuple correctly in rest of code.

Change-Id: Ief9b0c85e1ecc5e905ba50ae5e6d559b1c76b793
Reviewed-on: http://review.membase.org/5140
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Reviewed-by: Sean Lynch <seanlynch5@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>