Always use only a single ebucketmigrator process per src/dst
We also always use the same name for the connection. This allows the
connection to resume even when we intentionally interrupt it to change
the filters.

Since ebucketmigrator drops any messages for unasked for vbuckets on
the floor, we shouldn't need to worry about receiving any extra
mutations, but we do need to be careful that we don't miss any in
cases where we do want to backfill.

Change-Id: I5a0668a5b0f601d19cf2017dfd73da762e221ab8
Reviewed-on: http://review.membase.org/5221
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>