implemented robust misc:parallel_map
Next commit will replace misc:pmap with that new function. The problem
with pmap is that it leaves intermediate process alive if timeout
hits.

Another aspect of misc:parallel_map is that it carefully handles case
of calling processes being killed. And as pmap it makes sure no
messages are left in calling process mailbox.

Change-Id: I5cf041a58353ef61293174f22d24b0849cbc1292
Reviewed-on: http://review.membase.org/6108
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>