optimized config replication
Most significant change is in ns_node_disco_rep_events to queue config
replication instead of waiting for it's completion. This makes sure
that ns_config_events is not stuck on replication, so it's subscribers
are notified in time. This is slight semantical change (so it's not
purely optimization), but I'm sure it's worthy.

This leads to queueing of changes in ns_config_rep mailbox. Which we
also optimize a bit by squashing all available changes.

Change-Id: If1646d7e3c48e9b650e3085c18215fb2ad36b8e7
Reviewed-on: http://review.membase.org/6111
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Dale Harvey <daleharvey@arandomurl.com>
Tested-by: Dale Harvey <daleharvey@arandomurl.com>