Made aggregation of two stats samples faster
'Adding' two orddicts can be easily done with O(N) instead of
O(N²). Which is what we're doing here.

We're not using orddict:merge because our behavior is to only use keys
that are present in first proplist. So that we don't try to
merge/display stats that are unsupported/unkown on this node.

Badarith case is likely slightly slower (stats_reader may return
undefined in some rare cases), but it's very uncommon.

Change-Id: I0d52a2bac8bb8bc81cbc6fc87306cd5664415ff4
Reviewed-on: http://review.membase.org/6207
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Dale Harvey <daleharvey@arandomurl.com>
Reviewed-by: Dale Harvey <daleharvey@arandomurl.com>