get rid of null sample values in collector
This fixes badarith in menelaus_stats:basic_stats.

We're converting them to 0 anyways so it makes little sense.

The only place in stats that still produces null values is
samples_to_proplists which is final data transformation before sending
it to client. And 'null' is used to mark missing values, which may
happen after membase upgrade.

Change-Id: Ib9e1e94179fe6fc718e42f264b2bbb7ee880fe00
Reviewed-on: http://review.membase.org/6514
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>