simplified and fixed hostname port cutting logic
New logic is to strip port _only_ if all hostnames have default port
(8091).

So that if you have only single host with non-default port it's not
hidden from you. This was especially problematic with single active
host with port 9000 and single pending-add host with port 9001, in
which case old logic was stripping port confusing you.

Change-Id: I09f70703370f8020f61ea4d608d1758421544668
Reviewed-on: http://review.membase.org/6522
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>