batch replication changes during rebalance
So that we avoid too frequent re-connects to named tap streams and
re-starts of backfill.

The approach is to record replication changes and instead of applying
them immediately, apply them only if a) there's some period of
stability b) we need to apply at least one of them. Later is necessary
in case where we need to stop replication to new master before doing
takover into it. It is the only case where we need to sync data-safety
wise because a) we don't reorder anything (apply all or delay all) b)
only vbucket transfers can trigger bad thing (2 incoming streams for
one vbucket)

And if rebalance aborted (and we lose batched changes) then it's
janitor's work to setup replication.

Change-Id: I9866f22fa8d92a44b92b64bbd5f1632ec351c7a6
Reviewed-on: http://review.couchbase.org/6769
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>