Auto-failover mails mention node. MB-4028
The auto-failover alert mails do now mention the node
that was/couldn't be failovered.
othenodesdown mentions all nodes that are currently down.
 - auto_failover_node mentions the node that was failovered
 - auto_failover_other_nodes_down mentions all nodes that are
currently down.
 - auto_failover_cluster_too_small mentions the node that couldn't
be auto-failvored
 - auto_failvoer_maximum_reached doesn't mention any nodes, as it
is sent out only once. It would be misleading to mention a node on
the first time, but when another node goes not sending out any info.
The admin should just check what went wrong. Also, this is an
auto-failover alert and not a general "node down"-alert.

This commit fixes MB-4028.

Change-Id: I9df8c73d8b7c7be68f2813c9086b50aa2e804f32
Reviewed-on: http://review.couchbase.org/7743
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>