Use current settings for test email and display errors. MB-4029, MB-4031
The "Test email" button for the auto-failover email alerts
not uses the settings as set above, they don't need to be
(and won't be) saved.

In case the sending of the email didn't work, the button
will display "Error!" and will be disabled until the
settings got changed.

This commits also contains the removal of some log messages
that were there for debug purpose only.

This commit fixes MB-4029 and MB-4031.

Change-Id: Ida9218ceefea7fe4a2681592c2fed9c7a83afa70
Reviewed-on: http://review.couchbase.org/7745
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>