Add a possibility to save OTP cookie to the file system.
When it's needed to the caller 'cookiefile' environment variable can be set to
the ns_server. In that case whenever OTP cookie is changed it will be saved to
the path specified by the variable. In particular, this is needed to be able
to shutdown the server gracefully from outside. All the cookie management
machinery is now performed by separate gen_server (ns_cookie_manager). That is
needed since cookie synchronization can be triggered from different
processes. And now this can have visible side-effects which must be serialized.

Change-Id: Ie1b88d232cdba5938a813311d5a519ffda7a263a
Reviewed-on: http://review.couchbase.org/8015
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>