[bp] Add a possibility to save OTP cookie to the file system.
When it's needed to the caller 'cookiefile' environment variable can
be set to the ns_server. In that case whenever OTP cookie is changed
it will be saved to the path specified by the variable. In particular,
this is needed to be able to shutdown the server gracefully from
outside. All the cookie management machinery is now performed by
separate gen_server (ns_cookie_manager). That is needed since cookie
synchronization can be triggered from different processes. And now
this can have visible side-effects which must be serialized.

Change-Id: I85444fab0f9477540830a096a7c23859a8ab415c
Reviewed-on: http://review.couchbase.org/13432
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>