Display a warning when node was auto-failovered. MB-4074
When a node was auto-failovered, display a warning at the server
screen including a "reset quota" button.
The quota can also be reset on the settings screen. It will only
be displayed if there is something to reset.

When an error occurs a popup is shown.

This commit fixes MB-4074.

Change-Id: Id1e4d0d559d6563577ab9fb079bed6b8919d6b0d
[alk@tut.by: hide auto-failover warning by default]
Signed-off-by: Aliaksey Kandratsenka <alk@tut.by>
Reviewed-on: http://review.couchbase.org/8010
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>