Support two modes of execution in mb_master.
There is known problem regarding rebalance upgrade. VBucket mover spawns
single mover processes on remote nodes using spawn_link function. For this to
work ns_vbucket_mover modules must be absolutely the same on both nodes. This
means that ns_vbucket_mover can't be changed at all in order to preserve
compatibility. This commit is part of the work to resolve the issue. VBucket
mover is always launched on a master node. The idea is to ensure that master
node is always of the latest available version. In this case vbucket mover can
be changed in such a way that does not require spawning on remote nodes . But
to achieve desired master selection behavior it's insufficient to change only
new versions of mb_master. That's why this intermediate version
exists. Upgrades to this version from older ones work fine since vbucket mover
is unchanged. At the same time mb_master is made to support two modes of
execution: one that is used by older versions and one that will be used by
newer versions. The latter mode is enabled when the cluster contains only the
nodes that support it. And whenever some node of older version is added to a
cluster, behavior is reverted to the old one. Once cluster contains only new
nodes and operates in new mode, other 2.0 nodes can be connected to it. One of
those 2.0 nodes will take over the mastership and it will be possible to
perform rebalance upgrade even if ns_vbucket_mover will have changed.

Change-Id: I3a1b3c2c8413cb04f6641b69d39203280dc74e80
Reviewed-on: http://review.couchbase.org/9159
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>