Move the code moving single vbucket data into a separate module.
No anonymous or private functions are spawned on remote nodes
now. Additionally, since all the code performing the actual move is
concentrated in a single module, it will be easier to update it on
remote nodes. Together with new master election algorithm this will
allow us to perform cross-version rebalance even if vbucket mover
versions differ across the cluster.

Change-Id: I116a3969e2832caf7a8aadd63268a5e8d2311f36
Reviewed-on: http://review.couchbase.org/8577
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Dale Harvey <daleharvey@arandomurl.com>
Reviewed-by: Dale Harvey <daleharvey@arandomurl.com>