don't keep-alive http connections that trap exits
In join-related requests that can restart mochiweb we're trapping
exits so that shutdown of mochiweb doesn't close our connection
prematurely.

But with newer mochiweb it also leads to problems performing same
requests again when connection is reused (as httpc apparently
does). This likely happens because of 'EXIT' messages that are left on
process mailbox.

So we'll just close connection after such request is performed, which
is perfectly legal HTTP server behavior.

Change-Id: I13349dac6927db5eb894a25d4dd8f892e43c4854
Reviewed-on: http://review.couchbase.org/8948
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>