Add initial vclock to per node keys that changed since 1.6.5.
This change fixes MB-4219.

When config merging routines encounter two values not having vclocks
attached to them then the value that came from remote node is
chosen. This can cause certain races. Particularly this is the cause
of upgrade issues seen on windows: when node is started for the first
time it upgrades old config but then it pulls the config from other
nodes and prefers foreign value (outdated one) to its own. This change
introduces initial vclocks for such vulnerable keys so that new values
will be preferred.

Change-Id: Ibe85213eb8df2392f6b4b34413062c0aed19c0a1
Reviewed-on: http://review.couchbase.org/9206
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>