Adapt ns_log_sink to the messages without a code.
This allows us to send usual log messages to this sink so that the
user can see them too. Such messages are logged with a fake 0 code. On
the other hand they are still announced locally with undefined code
for subscribers to know for sure that there are no code attached to
the message.

Change-Id: Ibfb083fdee3cd981115c17246b9f21edc6a15e50
Reviewed-on: http://review.couchbase.org/9374
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>