[bp] Adapt ns_log_sink to the messages without a code.
This allows us to send usual log messages to this sink so that the
user can see them too. Such messages are logged with a fake 0 code. On
the other hand they are still announced locally with undefined code
for subscribers to know for sure that there are no code attached to
the message.

Originally reviewd on http://review.couchbase.org/9374

Change-Id: Ibe8d0e1e99d8885a3586c43e04ef9717676b1c48
Reviewed-on: http://review.couchbase.org/13535
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>