make sure autocompaction form validation is enabled at right time
Previously it was possible to enable it before form is actually
set. Because inside AutoCompactionSection#fillSettingsForm we have
call to Cell#getValue that may be satisfy it's body later.

So as a treatment we do it in continuation passing style where
formValidationEnabled cell is set in function that's called when
fillSettingsForm is done.

Change-Id: I4c399c52d7e595b1386fa033f0d7feceec468f7d
Reviewed-on: http://review.couchbase.org/9890
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Benjamin Young <benjamin@couchbase.com>