Fix CAPI's ensure_full_commit() method
ensure_full_commit() used to return 0 as the instance_start_time for a database
but as a binary string. The right data type to send it as is an integer as the
get_db_info() method does. This changeset fixes this inconsistency by having
ensure_full_commit() also return an integer. The problem was discovered during
XDCR integration testing.

Change-Id: I6eef64ef2eb00ece0bdf597389f5fe1766f2a56d
Reviewed-on: http://review.couchbase.org/10219
Tested-by: Srinivas Vadlamani <srinivas@couchbase.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>