Ignore non-XDC Type Rep Docs in Cancellation
Cancellation is triggered in response to a deleted doc but deleted
docs don't have a body, so we cannot determine the doc's type. So, in
order to trigger cancellation only for xdc type docs, we can simply
look up the ?XSTORE to check whether an entry for the doc exists. If
there isn't one, we simply ignore the deleted doc.

Change-Id: I4d33f499f4063c31211a90e3a8a5f0756bfba29e
Reviewed-on: http://review.couchbase.org/10459
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>