Store XDC Rep Info Even For Failed Attempts
Previously, the XDC replication manager used to store info of an XDC
rep doc in ?XSTORE only if replication was triggered successfully.
This is not the right strategy during addition of nodes to a running
cluster. Due to the asynchronous nature of bucket/vbucket creation,
immediately after a node is added, it may try to trigger replication
for an existing rep doc. However, if bucket creation isn't complete at
this time, then it should wait until it gets a config change update
from ns_config. In order to correctly react to a config change (the
maybe_adjust_all_xdc_replication() function is responsible for this),
the ?XSTORE must have a record of every XDC rep doc previously seen.

This change set makes sure that we add an entry to ?XSTORE even for
failed replication attempts so that they may be retried later. The
?XSTORE state is cleaned up upon deletion of a rep doc.

Change-Id: Icdbeee4a51ebc089e5642bb4e4491a8e92bbea3f
Reviewed-on: http://review.couchbase.org/10461
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>