Fix Incorrect Source/Target Bucket Name Parsing
For a push based couch replication of a local database to a remote
location, the source attribute of the #rep record is a binary data
type while the target is a tuple. This change set parses these
attributes correctly in the maybe_adjust_all_replications() and
maybe_retry_all_couch_replications() functions of the XDC replication
manager.

Change-Id: I832f4c147cec6fe4ba64a3c6f96977d83f2cb78e
Reviewed-on: http://review.couchbase.org/10464
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>