handle localhost when building replicator db link
When node's name is still ns_1@127.0.0.1 we cannot use hostname part
of node name to advertise it's address. Our usual approach is to use
address of the socket that's serving this request. So that
multi-interface host will advertise 'right' address depending on what
interface is reached.

This fixes XHR error when trying to create replication via UI.

Change-Id: I8c2a6d62a9842f3082791c2e3ac0caa232981955
Reviewed-on: http://review.couchbase.org/10634
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Srinivas Vadlamani <srinivas@couchbase.com>
Reviewed-by: Srinivas Vadlamani <srinivas@couchbase.com>