fixed 404 handling on pool details
Previous implementation was creating fake value with special marker
and then reacting on it in observer-function. But fake value did not
work with replications section (that's expecting links to
replication-related URLs). So I decided to just produce undefined in
missingValueProducer while triggering application reloading.

This fixes UI crashes in replications when cluster is reinitialized
and pools details request returns 404.

Change-Id: I2ac3607cfb41e19442827f8f5aee2a87a3f86984
Reviewed-on: http://review.couchbase.org/11139
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>