Retry couch replications in the same process
Previously, a separate timer process was run on a periodic basis to
retry failed replications. This can cause race conditions with the
main process and with other timer processes.

This changeset fixes the issue by having the timer process send
periodic messages to the main process which then actually retries the
failed replications.

Change-Id: Iabdd5fef20e2c7587edf15f9b52eda24457f4e94
Reviewed-on: http://review.couchbase.org/10828
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Srinivas Vadlamani <srinivas@couchbase.com>