have db path in couch config and restart world on change.MB-4466
2 issues caused exact failure in logs of MB-4466. First is that node
got old couchdb config key of ns_config replicated to it. That caused
cb_config_couch_sync to apply that setting and restart
couch. Restarting couch caused xdc_replication_manager to fail. And
starting it back failed, because xdc_replication_manager could not
contact couch that was in the middle of restarting. That was second
and direct cause of failure.

After tinkering with that issue I have decided to implement few
changes:

* we don't expose node's db and index path in ns_config. We keep it in
  (local) couch_config and replicate it via heartbeats for
  UI/diagnostics.

* we restart entire ns_server application when data & index path is
  changed. That ensures clean data & index path change. Note that
  implementation is not exactly application:{stop,start} sequence due
  to some group_leader issues.

* while changing this area I've simplified old and overengineered
  internal API designed for richer storage model.

* and also changing data & index path is not allowed anymore when
  there is at least one bucket defined. That didn't work in past
  anyways.

Change-Id: If5ebd9b1ecfa1d8e71f62b065948075de4eac674
Reviewed-on: http://review.couchbase.org/11238
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>