fixed data path validation and made it 1.7-compatible. MB-4465
First problem was broken validation where it was possible to screw
things up by POSTing nothing.

Second problem was that changing data path parameter from path to
db_path was not backwards-compatible, so we revert that and make
index_path be equal to data path if index_path is not in POST
parameters.

Change-Id: I657b1e21a66eda6ef9f5c01b2c50e2ffb48ef8b0
Reviewed-on: http://review.couchbase.org/11234
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>