Deliver only buckets changes having needed bucket config in them.
Sometimes during bucket deletion when capi_set_view_manager is not yet
terminated it gets a {bucket, ...} message that already does not
contain the config for the appropriate bucket. In this case
capi_set_view_manager crashes. This crash seems to be completely
harmless but just to avoid unnecessary noise in the log file from now
we will filter such messages out.

Change-Id: I865b56f7628bcf2f6d8a469c7d7b14108e52ec6f
Reviewed-on: http://review.couchbase.org/11545
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>