Keep a list of design documents in capi_set_view_manager.
Instead of fetching design docs from couch every time something
changes we now keep a list of known design documents. And now we
define view groups in a single place when we get notifications from
master db watcher. And we try to satisfy the invariant that there's a
defined set view group for every design document that is in this
list. But there are few cases when this invariant is broken:

  1. when the user deletes some document
  2. when the user updates some document.

In both of these cases capi_set_view_manager will eventually get a
notification about the change and will either delete the document from
the list (for the first case) or define a group again (for the second
case). But couch_set_view:set_partition_states (if run in between)
will throw either {not_found, deleted} or view_undefined errors
correspondingly. If this happens we just ignore these errors.

Change-Id: I066e26448cf7eb3cb05356d2407585d3fb41e5db
Reviewed-on: http://review.couchbase.org/11640
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>