handle compatible mode heartbeats in normal mode
In logs of MB-4507 we see mb_master crash when it receives compatible
mode heartbeat with versioning mark from last node. It switches to
normal mode, but node info is atom (not tuple) because node that sent
heartbeat is still in compatible mode.

We handle such hearbeat messages in same was as we do in compatible
mode (comparing node names). Every other case is handled in same way
as before.

Change-Id: Ied4abf9ef4451f1b7d7c9c2da27698b118040d6e
Reviewed-on: http://review.couchbase.org/11616
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>