Do not remove dead vbuckets from indexes right away.
If certain vbuckets has become dead but is still belongs to this node
according to vbucket map we will not remove it from indexes. But it
will be removed when vbucket map changes accordingly. This allows to
delay vbucket map change during rebalance safely. Additionally helps
to preserve indexes when server is restarted.

Change-Id: I2da5cfb86386c171c095efad43248a305d0752fe
Reviewed-on: http://review.couchbase.org/11749
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>