made janitor drop rebalance status iff really not running.MB-4579
maybe_stop_replication_status was assuming if janitor is run,
replication is not running. So we simply reset replication status.

But we in fact do call janitor as part of rebalance of each bucket. So
janitor was incorrectly resetting replication status to "not running".

Fix is to check in orchestrator (which should be local) if it's really
running rebalance or not.

Change-Id: I890a5f8739bae7be1da562a51e9ae4d5d40bf028
Reviewed-on: http://review.couchbase.org/11904
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>