refresh DAL.cells.capiBase value instead of keeping old value
Previous assumption was that capiBase don't need to change. So we just
kept old value forever, thinking that there's small benefit of that.

But this value _can_ change in fact. When node is started, it's
capiBase is 127.0.0.1:8902, when first other node joins it, we rename
it, thus changing capiBase.

This fixes bug when view creation via UI could fail sometimes.

Change-Id: I63e40b359c8ec19e6128613089d7420b1ed1a641
Reviewed-on: http://review.couchbase.org/11931
Reviewed-by: Dale Harvey <daleharvey@arandomurl.com>
Tested-by: Dale Harvey <daleharvey@arandomurl.com>