Set destination vbucket state to active early if master's missing.
Previously the state of new master for vbuckets that lacked old master
was set to active by janitor (which runs when rebalance is not in
progress). But it prevented capi_set_view_manager on destination from
adding the vbucket to indexes. But with recent changes
ns_vbucket_mover waits till all the vbuckets are added to
indexes. Thus rebalance would never complete.

Change-Id: I35941837ae9999a0471540a18d4530e4b3e7df3a
Reviewed-on: http://review.couchbase.org/12323
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Srinivas Vadlamani <srinivas@couchbase.com>