Split single_bucket_sup into two supervisors.
So now there is a separate supervisor for ns_memcached and all the
processes that depend on it. On the other hand capi_set_view_manager
and capi_ddoc_replication_srv are independent of each other and
ns_memcached_sup. Thus it's valid to use one_for_one restart
strategy. With this change some transient crash in, for instance,
capi_set_view_manager does not compel stats_reader and friends to
restart.

Change-Id: I32200f3bda241ddb2d6fa7999a106c9b75df9402
Reviewed-on: http://review.couchbase.org/12330
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>