[bp] initial replicator-on-destination implementation
We now run ebucketmigrator on destination node rather than
source. This will allow us to do quick failover because with pull
replication it's not necessary to wait for node being failed over to
stop push replication.

NOTE: this is not backwards-compatible with 1.8.0. This will be
addressed as part of future 1.8 work.

Originally reviewed on http://review.couchbase.org/12348

Conflicts:

	src/ns_janitor.erl

Change-Id: Ie06189d6e436aeced58f4c3d82b44968d53872a8
Reviewed-on: http://review.couchbase.org/13954
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>