initial replicator-on-destination implementation
We now run ebucketmigrator on destination node rather than
source. This will allow us to do quick failover because with pull
replication it's not necessary to wait for node being failed over to
stop push replication.

NOTE: this is not backwards-compatible with 1.8.0. This will be
addressed as part of future 1.8 work.

Change-Id: I9ac4f030618adce424f3e5aaa8441cd4d458db4d
Reviewed-on: http://review.couchbase.org/12348
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>