ns_pubsub:subscribe_link uses intermediate process.
With pure gen_event handler the subscriber must handle gen_event_EXIT
messages properly. And it's pretty easy to forget about this. In such
a case a crash inside a handler will be ignored. And depending on the
subscriber it can leave it in non-functional state. Thus we introduce
an intermediate process between subscriber and gen_event which handles
this. And if something wrong happens to a handler it exits with
non-normal reason. And this will also terminate the subscriber.

NOTE: if the subscriber sets trap_exit to true then it still needs to
handle EXIT messages from intermediate process.

Change-Id: I6b9c3fddaf387ea1f92048d8f240b57b24589ce9
Reviewed-on: http://review.couchbase.org/12421
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>