Set couch_replicator connnection timeout
The default timeout was too small and was causing the xdcr source to
retry requests too frequently. If the destination was already busy
servicing the original request, resending requests frequently would
make it worse.

The new value is 5 mins. Not sure if this is too conservative but if
it turns out to be so, it can be changed later or made an environment
variable.

Change-Id: I4b074ddd65f09afb8f43f936c8889620a06e0871
Reviewed-on: http://review.couchbase.org/13660
Tested-by: Srinivas Vadlamani <srinivas@couchbase.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>