disallow known wrong fields in setAutoCompaction POST
Recent change of API to inclide [size] and [percentage] broke some
testrunner tests that are using older API. Whats making it worse is
that by definition lack of xxxFragmentationThreshold field in POST
parameters is interperted as "no thanks". This caused testrunner to
disable autocompaction instead of setting it to default.

To make breakage like that visible we will signal error if old API use
is detected.

Unfortunately use of whitelist approach is hard here because
parse_validate_auto_compaction_settings is also invoked as part of
bucket create/update requests with rest of bucket parameters present
as well.

Change-Id: I478f8edc3563de9e2b8bc6bc079cb9e6ca4fc26d
Reviewed-on: http://review.couchbase.org/13428
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>