log normal and not scary message when retry_not_ready_vbuckets
Currently in rebalancing clusters with replica count more then 1 you
can quite often seen "crashes" which are not crashes in fact. But
voluntary death of ebucketmigrators.

This voluntarily death happens if some vbuckets we need to replicate
from are not yet backfilled. In such cases we do some progress with
ready vbuckets and terminate after 30 seconds. This commit changes
crash to normal exit and friendly info message.

Change-Id: Id6bc8d71e6cdcaaa77f02394caafeb1dc7887abd
Reviewed-on: http://review.couchbase.org/13799
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Srinivas Vadlamani <srinivas@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>