don't try to force local ddoc replication on janitor pass.MB-4825
The original idea was to ping ddoc replication srv that some vbuckets
may have been created. This was when we replicated ddocs to each
vbucket database.

It was broken even from day 0. Because it only touches local ddoc
replication srv. Even if this node doesn't run bucket yet/anymore.

Good thing is that it's not needed anymore. Because we don't replicate
ddocs to vbucket databases. So ddoc_replication_srv doesn't need any
notifications from janitor.

Change-Id: I759e457c984f94384eecfc673a7a0921111de99f
Reviewed-on: http://review.couchbase.org/13752
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>