try not creating bucket if it's running on any cluster nodes
We'll do best effort check if ns_memcached for this bucket still
exists anywhere and refuse creating it.

Best-effort here means that we'll try to reach other nodes, but with
sane timeout and if we cannot we'll assume it's fine. It is consistent
with current behavior and don't want to change that.

It is intended to prevent issues when somebody tries to re-create
bucket too soon after it was deleted. And to protect us from any
issues in ep-engine/bucket-engine that may delay bucket deletion. See
for example http://www.couchbase.com/issues/browse/MB-4329

Change-Id: I8d66cb5dc9603b4da8ad9ca430a0cd112e2a422f
Reviewed-on: http://review.couchbase.org/14182
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>