Don't use num_items_for_persistence during xdcr checkpointing
ep-engine flusher decrements num_items_for_persistence before
actually persisting an item, so relying on its value becoming zero
before updating the checkpoint document is not safe. Instead, we wait
for the current open checkpoint id to be persisted. If it gets
persisted before we timeout, the checkpoint document is updated.
Otherwise, we don't checkpoint.

Change-Id: I1e86be03a67d60f9a22a36d5ef052fae53b414e9
Reviewed-on: http://review.couchbase.org/13921
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Srinivas Vadlamani <srinivas@couchbase.com>