Explicitly trigger doc replication upon startup
Due to a race condition, it may so happen that after a new node is
added to a cluster, the design/replicator docs on existing nodes do
not replicated to the new node. This change set explicitly triggers
the replication of all existing docs by sending the
replicate_newnodes_docs message to all the existing nodes.

Change-Id: I3d3f47d3a2546334dfdb3b4052019e06b99c33f2
Reviewed-on: http://review.couchbase.org/13928
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Srinivas Vadlamani <srinivas@couchbase.com>