Choose between old and new replication styles.
cb_replication process (running on master node) keeps track of
replication modes for each bucket. Replication mode can be switched
only in ns_janitor during cleanup. The mode stored in cb_replication
is used only on master node (to avoid switches during rebalance). All
the other nodes determine which mode to use only based on
heartbeats. This means that certain lags are possible. But this seems
to be acceptable.

Change-Id: I36f22df51f722d73f32e17e8e63c02c76ddbb434
Reviewed-on: http://review.couchbase.org/13966
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>