Handle noproc while monitoring couch replications
There is a small delay between when we spawn a new Couch replication
and when we start monitoring it. If there is nothing to replicate,
the Couch replication process may finish before we even start
monitoring it and we will get the noproc error. This change set adds
explicit handling for this case by considering such a replication as
"completed".

Change-Id: I5f9235ea671f53ce9351cf8d0bc864ce1ad129f0
Reviewed-on: http://review.couchbase.org/14373
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Srinivas Vadlamani <srinivas@couchbase.com>