backfill all new replicas and master before doing vbucket takover
It has effect of limiting backfills on source and destination
side. But most important reason for doing this is that by switching to
built replicas we have more reliable rebalance. It should be possible
to stop it at any time and things will be in much safer shape now
compared to previous implementation.

Change-Id: I6c9c8aef6197178558c61940598e4aa43715df0c
Reviewed-on: http://review.couchbase.org/14193
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>