handle very quick completion of vbucket mover.MB-4964
With empty vbucket movements optimization we're now hitting old race.

Particularly we're starting monitoring child mover after it is
started. And when it completes too quickly we get noproc instead of
it's real exit reason.

Because entire rebalancer process is not trapping exits getting noproc
means that exit was normal, otherwise rebalancer would be dead before
receiving this.

Change-Id: I2a682b18cd51306326d644fcf59bbd9f4c0aec64
Reviewed-on: http://review.couchbase.org/14361
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Srinivas Vadlamani <srinivas@couchbase.com>