MB-5033 Handle messages with undefined code in ns_log correctly.
If some module doesn't define ns_log_cat message and logs something to
ns_log_sink, then an undefined code will be assigned to this
message. This undefined code is changed to 0 just before
logging. Previously it was done right in ns_log:handle_cast (do_log
clause). But this is not correct since do_log messages are abcasted to
all the nodes. Some of them can be of old version. On old nodes
messages with undefined codes will be just ignored when shown in the
UI (because of the catch around corresponding io_lib:format). And they
will crash diag_handler because of no catch around io_lib:format used
there.

Change-Id: I3ebeb0b2d0e434eeb0d55689d74748ccfd7a767f
Reviewed-on: http://review.couchbase.org/14723
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>