MB-5008 Retry net_kernel:start several times on duplicate name.
epmd unregisters old node names when a socket that was used to
register this name is closed. This is what happens when
net_kernel:stop is called. But there's no guarantee that subsequent
register name request (issued by net_kernel:start) will be served
after the old name has been unregistered. Because of this we can get
{error, duplicate_name} on net_kernel:start (even though it seems to
be quite unlikely). With this change we will retry net_kernel:start
several times if we get this duplicate_name error.

Change-Id: I2fab246897dcff1fcf73cebc65aef6ff4247286f
Reviewed-on: http://review.couchbase.org/14728
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>