Include debug log file into diag by default.
The same holds for /sasl_logs and different flavors of cbbrowse_logs.

The reasoning is as follows. It's better to have more than less
information by default. And also prevents cases when one just forgets
to specify that debug info is required, but when it's figured out,
the state of the cluster has already changed significantly.

Change-Id: I85c6924cbaa27cf45d2906c220a1f3834973280b
Reviewed-on: http://review.couchbase.org/14729
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>