complete chain switch in janitor if new master is active.MB-4988
When rebalance is stopped and future master is already active
previously janitor would use new master but mess up replicas.

We're conservatively detect if active vbucket is on future master as
seen on fast forward map. And if all nodes in fast forward chain are
replicas. In that case we can safely assume takeover was done and we
simply update vbucket map.

Change-Id: Id268b701c548277d47a765f788033cecd09840ac
Reviewed-on: http://review.couchbase.org/14758
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>