delete all node's unused bucket db files before rebalance.MB-5115
As noted in MB-5115 we were not treating partially failed over node as
'new' when rebalancing back in, which could lead to failure to delete
old bucket's db files.

Correct fix seems to be to get rid of _any_ unused db files on all
nodes at the beginning of rebalance.

Change-Id: Ic1e5af53063f34d6791c238669461cd690ac6a3f
Reviewed-on: http://review.couchbase.org/15051
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>