reimplemented backwards-compat for change_vbucket_filter. MB-5108
Previously I assumed sending erl_eval closures to old nodes works. But
we've found this assumption to not hold.

So new approach (because we still care about latency in fast path) is
to send call to have_local_change_vbucket_filter and
local_change_vbucket_filter in parallel and then use results of both
to reliably detect if we're talking with old node which only support
old-style (unreliable) vbucket filter change.

Change-Id: I7394f04996b0ebd8c717cc8662181914bb8871c1
Reviewed-on: http://review.couchbase.org/15366
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>