MB-5307: unregister tap name even of not vbuckets are ready
Previously it was possible that tap is not supposed to be reused _but_
has no vbuckets ready. So deregister_tap_client would not be sent,
then vbucket filter change arrives and we are supposedly reusing tap,
but intention is to reuse tap from previous instance and not from some
unknown source.

Change-Id: Ie590973193bfb152f7978395b502637efdf460d8
Reviewed-on: http://review.couchbase.org/16173
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>